feat(crypto): verify sig
